requirements.txt
# langchain==0.0.184
# PyPDF2==3.0.1
python-dotenv==1.0.0
# streamlit==1.18.1
openai==0.27.6
# faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
streamlit==1.27.0
langchain-google-genai==0.0.5
langchain-community==0.0.13
langchain-core==0.1.12
langchain==0.1.0
google-generativeai==0.3.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-pptx==0.6.21
python-docx==0.8.11
langchain-huggingface==0.0.6
# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2